Merge: Req4-5-2 to dev

See merge request metleeha/webmobile-final!10

+++ src/views/PostViewPage.vue
@@ -14,7 +14,7 @@
     <!-- POST Detail View -->
     <v-container v-if="loading">
       <!-- Edit Button -->
-        <v-layout xs12 justify-center class="post-edit">
+        <v-layout xs12 justify-center class="post-edit" v-if="authCheck()">
           <v-flex xs12 text-xs-center mt-5>
             <v-btn @click="newToggle = !newToggle" block flat large :color="newToggle? '#f0bebe':'#3a718c'"><v-icon left>fa-chevron-down</v-icon>Edit Post</v-btn>
           </v-flex>
@@ -41,7 +41,7 @@
       <!-- Created: Date, Author -->
       <v-layout row class="post-author">
         <v-flex xs12 justify-center>
-          <p>by admin, {{ date }}</p>
+          <p>by {{name}} - ({{ date }})</p>
         </v-flex>
       </v-layout>
       <!-- Post Body --> 
@@ -51,8 +51,8 @@
         </v-flex>
         <!-- btns -->
         <v-flex xs12 row class="post-btn">
-          <v-btn large outline color="red" @click="deletePost()">DELETE</v-btn>
           <v-btn large outline color="green" @click="goToPost()">Back</v-btn>
+          <v-btn v-if="authCheck()" large outline color="red" @click="deletePost()">DELETE</v-btn>
         </v-flex>
       </v-layout>
       <!--Disqus -->
@@ -82,6 +82,7 @@ export default {
 	data(){
     return{
       newToggle: false,
+      name: '누구더라...',
 		  date: '언제더라...',
 		  title: '로딩중...',
       body: '뭐더라...',
@@ -98,10 +99,12 @@ export default {
     },
   methods: {
     async getPostData(){
+      this.loading = false
       const pf = await FirebaseService.getPost(this.id)
       this.$store.state.postDetail = pf
     },
     setPostData(post) {
+      this.name = post.username
       this.title = post.title
       this.body = post.body
       var fulldate = new Date(post.created_at.seconds*1000)
@@ -109,9 +112,16 @@ export default {
       this.date = fulldate.toString().substring(4, 15)
       this.loading = true
     },
-    deletePost(){
-      FirebaseService.deletePost(this.id)
-      this.$router.push({name: 'post'})
+    async deletePost(){
+      // 해당 ID를 가진 portfolio를 DB에서 가져옴
+      const portfoliodata = await FirebaseService.getPortfolio(this.id)
+      // 접속한 유저 정보 DB에서 받아옴
+      const user = await FirebaseService.getUserDataAuth()
+      // 접속 유저정보와 DB 데이터비교 권한 인증
+      if(FirebaseService.authUserAndDB(portfoliodata, user)){
+        FirebaseService.deletePost(this.id)
+        this.goToPost()
+      }
     },
     updatePost(state){
       if(state){
@@ -122,6 +132,18 @@ export default {
     },
     goToPost(){
       this.$router.push({name: 'post'})
+    },
+    authCheck() {
+        const user = this.$store.getters.getUser;
+        if (user) {
+            if (user.authority == "master") {
+                return true;
+            } else if(user.authority == "member" && user.name == this.name) {
+                return true;
+            }
+        } else {
+            return false;
+        }
     }
   },
   computed: {
